@@ -52,7 +52,7 @@ return [
     |
     */
 
-    'url' => env('APP_URL', 'http://localhost/visaportal/'),
+    'url' => env('APP_URL', 'http://localhost/visaportal'),
 
     'asset_url' => env('ASSET_URL', 'http://localhost/visaportal/public'),
 
